@@ -1,9 +1,9 @@
 package mykotlinpackage.model
 
 import com.github.manosbatsis.scrudbeans.api.annotation.model.ScrudBean
-import javax.persistence.Entity
-import javax.persistence.Id
-import javax.persistence.IdClass
+import jakarta.persistence.Entity
+import jakarta.persistence.Id
+import jakarta.persistence.IdClass
 
 @ScrudBean
 @Entity
