@@ -2,9 +2,9 @@ package mykotlinpackage.model
 
 import io.swagger.v3.oas.annotations.media.Schema
 import java.io.Serializable
-import javax.persistence.Embeddable
-import javax.persistence.JoinColumn
-import javax.persistence.ManyToOne
+import jakarta.persistence.Embeddable
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.ManyToOne
 
 /** Embeddable many2many identifier for [Product] relationships */
 @Embeddable
